Focus first option if options change and is open.
Handle case where the options change (e.g. due to input filtering) and
the current focused index is not valid/appropriate.

For now just always reset to first focused option. In future, might want
to try and maintain the focus on the item that was already focused if
still in the list.