Fix inconsistent text rendering in input.
Ensure the input element matches the specified font for the control
so that there is no difference between displayed value and typed text
rendering.

Also, prevent the input computing a different size due to browser
default styles (e.g. margin, padding). Force a line-height of normal for
the control as Firefox prevents setting a smaller line-height for inputs
which can cause an discrepancy in how the value text is rendered vs
the input text.