Support options generation via function callback.
Reduce need for callers to manually setup a select for dynamic update
of options based on input value by allowing passed options prop to be a
function. The function receives changes to the input value and should
return the updated options to use. In addition, set readonly automatically
to false when a function for options is passed.

This makes creating a dynamic select as simple as passing options a
function like:
```
(inputValue: string) =>
  ["apple", "banana", "pear", "pineapple", "kiwi"].filter((fruit) =>
    fruit.startsWith(inputValue)
  );
```