Support remove handle for values.
Display a handle on multi values that can be clicked to remove the entry
from the values list of a multiple select.

Note: due to use of non-delegated event listeners in container ref
helper, use the `on:click` form for a corresponding non-delegated
listener on the button. Without this, attempting to prevent propagation
would fail and the control would toggle open on remove of value.
See also: https://github.com/solidjs/solid/issues/804#issuecomment-1027269781