Support disabling individual options in the list.
Prevent disabled options from being picked, but still display them in
the option list (with some styling to differentiate). By default,
consider no options as disabled. Support passing a custom
`isOptionDisabled` function to either `createSelect` or the `Select`
component to customise how an option is determined as disabled or not:

```jsx
<Select
  options={["apple", "pear", "kiwi"]}
  isOptionDisabled={option => option === "pear"}
/>
```

As part of this, also fix the 'click' handler on the container to ignore
clicks on any child of the list rather than just the list itself. Without
this fix, clicking a disabled option would still incorrectly close the
list.