React to `initialValue` even if initially undefined on `Select`.
Move the check for `undefined` into the effect itself. This enables the
`initialValue` to be changed from the default (`undefined`) later and
still have the `Select` component react to that.

Note this is done only for the `Select` component as `undefined` is
not a valid value to pass to `setValue` in the core. Supporting it would
introduce complexity because a `multiple` value should always be an
Array.