Support custom formatting in createOptions.
Pass a 'format' function to `createOptions` to
control formatting of option and value labels, a
'text' conversion of the value for comparison to
other options, as well as ability to customise
'Create' prefix when creating new options.

In addition, the highlight value from default
filtering is passed as metadata to the format
function to allow mixing the highlighted element
with other elements (such as maintaining an icon
even when filtering).

For #58 and could also address #57 and #56