Move existene check to createable callback.
Enable createable callback to determine whether to
have a create option added or not. Pass it a
default 'exists' boolean based on text comparison
for ease, but allow it to do any check by also
passing the existing options array. If the
callback returns undefined then don't add a create
option.

To maintain backwards compatibility, try to detect
if the callback accepts the new number of
parameters. If it doesn't, explicitly do the
exists check internally first and only call the
callback if no option already exists. Add a
warning to encourage consumers to update their
logic.

For #19