Separate text extraction from format.
They serve different purposes, so keep them
separate for clarity. In addition, this simplifies
the function signature and avoids the typing
nightmare of conditional return types (that isn't
actually properly supported today by Typescript).

As part of this change, standadise the handling of
'key' so that custom 'format', 'disable', and
'extractText' functions receive the keyed value
when 'key' is also specified. This feels most
intuitive. For those that need the full object,
they can always not pass 'key' and manage
explicitly.